chore: update for ngff refactoring
